sagemaker-core/pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "sagemaker-core"
dynamic = ["version"]
description = "An python package for sagemaker core functionalities"
authors = [
  {name = "AWS", email = "sagemaker-interests@amazon.com"}
]
readme = "README.rst"
dependencies = [
  # Add your dependencies here (Include lower and upper bounds as applicable)
    "boto3>=1.35.75,<2.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "PyYAML>=6.0, <7.0",
    "jsonschema<5.0.0",
    "platformdirs>=4.0.0, <5.0.0",
    "rich>=13.0.0, <14.0.0",
    "mock>4.0, <5.0",
    "importlib-metadata<7.0,>=1.4.0",
    "typing_extensions>=4.9.0",
    "pytz>=2021.1",
    "requests>=2.20.0,<3.0.0",
    "attrs>=20.3.0",
    "packaging>=20.0",
    "protobuf>=3.12,<5.0",
    "pandas>=1.0.0",
    "numpy>=1.9.0",
    # Dependencies migrated from sagemaker_utils
    "smdebug_rulesconfig>=1.0.1",
    "schema>=0.7.5",
    "omegaconf>=2.1.0",
    "torch>=1.9.0",
    "scipy>=1.5.0",
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.optional-dependencies]
codegen = [
    "black>=24.3.0, <25.0.0",
    "pandas>=2.0.0, <3.0.0",
    "pytest>=8.0.0, <9.0.0",
    "pylint>=3.0.0, <4.0.0"
]
test = [
    "pytest>=8.0.0, <9.0.0",
    "pytest-cov>=4.0.0",
    "pytest-xdist>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/aws/sagemaker-python-sdk"

[tool.setuptools.packages.find]
where = ["src"]
include = ["sagemaker*"]
namespaces = true

[tool.setuptools.package-data]
"sagemaker.core" = ["image_uri_config/*.json", "jumpstart/region_config.json"]

[tool.black]
line-length = 100
exclude = '\.ipynb$'

[tool.setuptools.dynamic]
version = { attr = "sagemaker.core._version.__version__"}

[tool.coverage.run]
omit = [
    "*/sagemaker/core/tools/resources_codegen.py"
]
